moved utils and algorithm to common spot in coreir instead of simulator
