+++ include/coreir/common/algorithm.h
