+++ include/coreir/common/utils.h
@@ -7,7 +7,7 @@
 #include "coreir/ir/types.h"
 #include "coreir/ir/wireable.h"
 
-#include "coreir/simulator/algorithm.h"
+#include "coreir/common/algorithm.h"
 
 namespace CoreIR {
