+++ src/simulator/op_graph.cpp
@@ -1,7 +1,7 @@
 #include "coreir/simulator/op_graph.h"
 
-#include "coreir/simulator/algorithm.h"
-#include "coreir/simulator/utils.h"
+#include "coreir/common/algorithm.h"
+#include "coreir/common/utils.h"
 
 using namespace std;
