+++ src/simulator/subcircuit.cpp
@@ -4,10 +4,10 @@
 #include "coreir/ir/types.h"
 #include "coreir/passes/transform/deletedeadinstances.h"
 #include "coreir/passes/transform/fold_constants.h"
-#include "coreir/simulator/algorithm.h"
 #include "coreir/simulator/op_graph.h"
 #include "coreir/simulator/wiring_utils.h"
-#include "coreir/simulator/utils.h"
+#include "coreir/common/algorithm.h"
+#include "coreir/common/utils.h"
 
 using namespace std;
